requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
cloudinary==1.34.0
dj-database-url==2.0.0
Django==3.2.20
django-ckeditor==6.3.2
django-js-asset==2.0.0
django-taggit==4.0.0
idna==3.4
Pillow==9.5.0
psycopg2-binary==2.9.5
python-memcached==1.59
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.7.1
urllib3==1.26.7
whitenoise==6.4.0
gunicorn==20.1.0
